requirements.txt
discord.py==2.2.3
openai_whisper==20230314
python-dotenv==1.0.0
